Move App\User to App\Models\User
