+++ database/factories/UserFactory.php
@@ -1,8 +1,8 @@
 <?php
 
-use App\User;
-use Illuminate\Support\Str;
+use App\Models\User;
 use Faker\Generator as Faker;
+use Illuminate\Support\Str;
 
 /*
 |--------------------------------------------------------------------------
